# Perf and other improvements for querying extensions